requirements.txt
aiofiles
aiogram
aiohttp[speedups]
asyncio-periodic
asyncpg
matplotlib
pillow
pycairo
